+++ src/Holding/index.js
@@ -2,3 +2,4 @@ export * from './CreateHolding';
 export * from './DuplicateHolding';
 export * from './EditHolding';
 export { default as HoldingsMarcContainer } from './HoldingsMarc';
+export * from './HoldingVersionHistory';