+++ src/Instance/InstanceDetails/InstanceDetails.test.js
@@ -1,7 +1,12 @@
-import React from 'react';
+import React, { act } from 'react';
 import '../../../test/jest/__mock__';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { screen, fireEvent } from '@folio/jest-config-stripes/testing-library/react';
+import {
+  screen,
+  fireEvent,
+  within,
+} from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
 import { DataContext } from '../../contexts';
 import { renderWithIntl, translationsProperties } from '../../../test/jest/helpers';
@@ -35,6 +40,7 @@ const instance = {
   notes: [],
   staffSuppress: false,
   discoverySuppress: false,
+  deleted: false,
   shared: false,
 };
 
@@ -51,7 +57,7 @@ const mockReferenceData = {
 
 const queryClient = new QueryClient();
 
-const actionMenu = jest.fn();
+const actionMenu = jest.fn(() => <button type="button">Action menu</button>);
 const onClose = jest.fn();
 const tagsEnabled = true;
 
@@ -143,6 +149,20 @@ describe('InstanceDetails', () => {
 
     expect(screen.getByText('Warning: Instance is marked suppressed from discovery and staff suppressed')).toBeInTheDocument();
   });
+  it('should show a correct Warning message banner when set for deletion, staff and discovery suppressed', () => {
+    const deletedAndBothSuppressedInstance = {
+      ...instance,
+      staffSuppress: true,
+      discoverySuppress: true,
+      deleted: true,
+    };
+    renderInstanceDetails({ instance: deletedAndBothSuppressedInstance });
+
+    const warningBanner = screen.getByText('Warning:', { exact: false });
+
+    expect(warningBanner.textContent).toEqual('Warning: Instance is set for deletion, suppressed from discovery, and staff suppressed');
+    expect(screen.getByText('Set for deletion')).toBeInTheDocument();
+  });
 
   it('expands and collapses the accordion sections', () => {
     renderInstanceDetails();
@@ -204,6 +224,56 @@ describe('InstanceDetails', () => {
       renderInstanceDetails();
 
       expect(screen.queryByRole('button', { name: 'Consortial holdings' })).not.toBeInTheDocument();
+    });
+  });
+
+  describe('Version history component', () => {
+    let versionHistoryButton;
+
+    beforeEach(async () => {
+      await act(async () => { renderInstanceDetails(); });
+
+      versionHistoryButton = screen.getByRole('button', { name: /version history/i });
+    });
+
+    it('should render version history button', async () => {
+      expect(versionHistoryButton).toBeInTheDocument();
+    });
+
+    describe('when click the button', () => {
+      beforeEach(async () => {
+        await act(() => userEvent.click(versionHistoryButton));
+      });
+
+      it('should render version history pane', () => {
+        expect(screen.getByRole('region', { name: /version history/i })).toBeInTheDocument();
+      });
+
+      it('should hide action menu', () => {
+        expect(screen.queryByText('Action menu')).not.toBeInTheDocument();
+      });
+
+      it('should disable tags button', () => {
+        expect(screen.getByRole('button', { name: /show tags/i })).toBeDisabled();
+      });
+
+      it('should disable version history button', () => {
+        expect(screen.getByRole('button', { name: /version history/i })).toBeDisabled();
+      });
+    });
+
+    describe('when click the close button', () => {
+      it('should hide the pane', async () => {
+        await act(() => userEvent.click(versionHistoryButton));
+
+        const versionHistoryPane = await screen.findByRole('region', { name: /version history/i });
+        expect(versionHistoryPane).toBeInTheDocument();
+
+        const closeButton = await within(versionHistoryPane).findByRole('button', { name: /close/i });
+        await act(() => userEvent.click(closeButton));
+
+        expect(screen.queryByRole('region', { name: /version history/i })).not.toBeInTheDocument();
+      });
     });
   });
 });