+++ src/components/BrowseResultsList/utils.js
@@ -5,6 +5,7 @@ import {
   browseClassificationOptions,
   browseModeOptions,
   browseClassificationIndexToId,
+  browseCallNumberIndexToId,
   FACETS,
   queryIndexes,
 } from '@folio/stripes-inventory-components';
@@ -21,7 +22,6 @@ export const isRowPreventsClick = (row, browseOption) => {
   const isItemHasNoRecords = row.totalRecords === 0;
 
   return isItemHasNoRecords || (
-    (browseOption === browseModeOptions.CALL_NUMBERS && !row.shelfKey) ||
     (browseOption === browseModeOptions.CONTRIBUTORS && !row.contributorNameTypeId) ||
     (browseOption === browseModeOptions.SUBJECTS && !row.totalRecords)
   );
@@ -97,9 +97,51 @@ const getClassificationQuery = (qindex, data, row) => {
   return query;
 };
 
+export const getFullCallNumber = (row) => {
+  const fullCallNumber = [row.callNumberPrefix, row.callNumber, row.callNumberSuffix].filter(Boolean).join(' ');
+
+  return fullCallNumber;
+};
+
+/**
+ * Constructs a query string for browsing call numbers based on the provided parameters.
+ *
+ * @param {string} qindex - The query index indicating the type of browse (e.g., callNumbers, dewey, lc).
+ * @param {Object} data - The data object containing configuration for call number browsing.
+ * @param {Object} row - The row object containing details of the item being browsed.
+ * @returns {string} - The constructed query string for browsing call numbers.
+ */
+const getCallNumberQuery = (qindex, data, row) => {
+  const fullCallNumber = getFullCallNumber(row);
+
+  const isCallNumberBrowse = Object.values(browseCallNumberOptions).includes(qindex);
+
+  if (!isCallNumberBrowse) {
+    return '';
+  }
+
+  let query = `itemFullCallNumbers="${fullCallNumber}"`;
+
+  const callNumberBrowseConfigId = browseCallNumberIndexToId[qindex];
+
+  const callNumberBrowseTypes = data?.callNumberBrowseConfig
+    .find(config => config.id === callNumberBrowseConfigId)?.typeIds;
+
+  const callNumberBrowseTypesQuery = callNumberBrowseTypes
+    ?.map(typeId => `item.effectiveCallNumberComponents.typeId=="${typeId}"`)
+    .join(' or ');
+
+  if (callNumberBrowseTypesQuery) {
+    query += ` and (${callNumberBrowseTypesQuery})`;
+  }
+
+  return query;
+};
+
 export const getSearchParams = (row, qindex, allFilters, data) => {
   const filters = getExtraFilters(row, qindex, allFilters);
   const classificationQuery = getClassificationQuery(qindex, data, row);
+  const callNumberQuery = getCallNumberQuery(qindex, data, row);
 
   const classificationOption = {
     qindex: queryIndexes.QUERY_SEARCH,
@@ -107,42 +149,20 @@ export const getSearchParams = (row, qindex, allFilters, data) => {
     ...filters,
   };
 
+  const callNumberOption = {
+    qindex: queryIndexes.QUERY_SEARCH,
+    query: callNumberQuery,
+    ...filters,
+  };
+
   const optionsMap = {
-    [browseModeOptions.CALL_NUMBERS]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.DEWEY]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.LIBRARY_OF_CONGRESS]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.LOCAL]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.NATIONAL_LIBRARY_OF_MEDICINE]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.OTHER]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
-    [browseModeOptions.SUPERINTENDENT]: {
-      qindex: queryIndexes.CALL_NUMBER,
-      query: row.shelfKey,
-      ...filters,
-    },
+    [browseModeOptions.CALL_NUMBERS]: callNumberOption,
+    [browseModeOptions.DEWEY]: callNumberOption,
+    [browseModeOptions.LIBRARY_OF_CONGRESS]: callNumberOption,
+    [browseModeOptions.LOCAL]: callNumberOption,
+    [browseModeOptions.NATIONAL_LIBRARY_OF_MEDICINE]: callNumberOption,
+    [browseModeOptions.OTHER]: callNumberOption,
+    [browseModeOptions.SUPERINTENDENT]: callNumberOption,
     [browseModeOptions.CLASSIFICATION_ALL]: classificationOption,
     [browseModeOptions.DEWEY_CLASSIFICATION]: classificationOption,
     [browseModeOptions.LC_CLASSIFICATION]: classificationOption,