+++ src/routes/ViewRequestsRoute.js
@@ -101,7 +101,7 @@ ViewRequestsRoute.manifest = {
     fetch: true,
     throwErrors: false,
     params: {
-      query: 'instanceId==:{id}',
+      query: 'instanceId==:{id} sortBy effectiveLocation.name callNumberPrefix callNumber callNumberSuffix',
       limit: '1000',
     },
   },